cloned_tasks/NexaTest---AI_Text_PreProcessing_Pipeline/requirements.txt
streamlit
numpy
pandas
scikit-learn
gensim
pyyaml
psycopg2-binary
sqlalchemy
plotly
matplotlib
pytest
chardet
gensim